コード整理とか
